refactor(jans-cedarling): organize structs into their own files
- rename AgamaPolicyStore to PolicySource
- move PolicySouceJson into it's own file
- move PolicySouceYaml into it's own file
